Refactoring of expose() methods
- Handle up to seven parameters.
- Handle const and non-const member methods, functions and lambdas
  (which need to be wrapped using wrap() method, so compiler can
  correctly deduce their type).
- Single base implementation for all variants of expose() method.
- Automatic conversion of Connection_options parameters.
- Automatic conversion of mysqlshdk::utils::nullable<T> parameters
  (such parameter will accept null in addition to any other types
  accepted by T).
- Only one specialization of Type_info template per type is necessary
  (no need for specializations for T, T&, const T&, etc.).
- Specializations of Type_info template moved to separate header files,
  classes calling expose() need to include them directly:
  - mysqlshdk/include/scripting/type_info/custom.h - custom Shell types,
  - mysqlshdk/include/scripting/type_info/generic.h - basic C++ types.

Additional changes
- Removed Cpp_object_bridge::add_varargs_method(), code using it was
  refactored to use expose() instead.
- Refactored all scripting methods which accept Connection_options
  parameter, they are now exported using expose().
- The member methods of Python's mysql/mysqlx modules are now wrapped as
  methods. Previously they were wrapped as functions, which caused that
  the full context (module_name.method_name) was not included in the
  thrown exceptions.
- Removed support for named parameters in Python functions, a proper
  solution needs to be implemented.

+++ unittest/test_utils/mod_testutils.cc
@@ -54,6 +54,10 @@
 #endif
 
 #include "mysqlshdk/libs/utils/debug.h"
+#include "mysqlshdk/include/scripting/obj_date.h"
+#include "mysqlshdk/include/scripting/type_info/custom.h"
+#include "mysqlshdk/include/scripting/type_info/generic.h"
+#include "mysqlshdk/include/shellcore/shell_init.h"
 #include "mysqlshdk/include/shellcore/utils_help.h"
 #include "mysqlshdk/libs/config/config_file.h"
 #include "mysqlshdk/libs/db/mysql/session.h"
@@ -80,8 +84,6 @@
 #include "modules/adminapi/common/metadata_management_mysql.h"
 #include "modules/mod_utils.h"
 #include "mysqlshdk/shellcore/shell_console.h"
-#include "mysqlshdk/include/shellcore/shell_init.h"
-#include "scripting/obj_date.h"
 
 // clang-format off
 #ifndef _WIN32
@@ -190,7 +192,8 @@ Testutils::Testutils(const std::string &sandbox_dir, bool dummy_mode,
   expose("skip", &Testutils::skip, "reason");
   expose("versionCheck", &Testutils::version_check, "v1", "op", "v2");
   expose("waitForDelayedGRStart", &Testutils::wait_for_delayed_gr_start, "port",
-         "rootpass", "?timeout", k_wait_delayed_gr_start_timeout);
+         "rootpass", "?timeout",
+         static_cast<int>(k_wait_delayed_gr_start_timeout));
   expose("mkdir", &Testutils::mk_dir, "name", "?recursive");
   expose("rmdir", &Testutils::rm_dir, "name", "?recursive");
   expose("chmod", &Testutils::ch_mod, "path", "mode");
